+++ components/org.wso2.carbon.identity.oauth2.grant.organizationswitch/pom.xml
@@ -20,14 +20,14 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 
     <parent>
-        <groupId>org.wso2.carbon.extension.identity.oauth2.grantType.organization.switch</groupId>
+        <groupId>org.wso2.carbon.extension.identity.oauth2.grantType.organizationswitch</groupId>
         <artifactId>identity-inbound-oauth2-grant-organization-switch</artifactId>
         <version>1.0.0-SNAPSHOT</version>
         <relativePath>../../pom.xml</relativePath>
     </parent>
 
     <modelVersion>4.0.0</modelVersion>
-    <artifactId>org.wso2.carbon.identity.oauth2.grant.organization.switch</artifactId>
+    <artifactId>org.wso2.carbon.identity.oauth2.grant.organizationswitch</artifactId>
     <name>WSO2 - Organization Switch Grant Type</name>
     <packaging>bundle</packaging>
 
@@ -73,11 +73,12 @@
                         <Bundle-SymbolicName>${project.artifactId}</Bundle-SymbolicName>
                         <Bundle-Name>${project.artifactId}</Bundle-Name>
                         <Bundle-Description>Organization Switch Grant Bundle</Bundle-Description>
-                        <Private-Package>org.wso2.carbon.identity.organization.management.oauth2.grant.internal
+                        <Private-Package>
+                            org.wso2.carbon.identity.oauth2.grant.organizationswitch.internal
                         </Private-Package>
                         <Export-Package>
-                            !org.wso2.carbon.identity.organization.management.oauth2.grant.internal,
-                            org.wso2.carbon.identity.organization.management.oauth2.grant.*; version="${project.version}"
+                            !org.wso2.carbon.identity.oauth2.grant.organizationswitch.internal,
+                            org.wso2.carbon.identity.oauth2.grant.organizationswitch.*; version="${project.version}"
                         </Export-Package>
                         <Import-Package>
                             org.apache.commons.lang;version="${org.apache.commons.lang.imp.pkg.version.range}",
