+++ pom.xml
@@ -16,25 +16,27 @@
  ~ specific language governing permissions and limitations
  ~ under the License.
 	 -->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>org.wso2</groupId>
         <artifactId>wso2</artifactId>
         <version>1.4</version>
     </parent>
-    <groupId>org.wso2.carbon.extension.identity.oauth2.grantType.organization.switch</groupId>
+    <groupId>org.wso2.carbon.extension.identity.oauth2.grantType.organizationswitch</groupId>
     <artifactId>identity-inbound-oauth2-grant-organization-switch</artifactId>
     <version>1.0.0-SNAPSHOT</version>
     <packaging>pom</packaging>
     <name>WSO2 Carbon Extension - Organization Switch Grant Type</name>
     <url>http://wso2.org</url>
     <modules>
-        <module>components/org.wso2.carbon.identity.oauth2.grant.organization.switch</module>
+        <module>components/org.wso2.carbon.identity.oauth2.grant.organizationswitch</module>
     </modules>
 
     <scm>
-        <connection>scm:git:https://github.com/wso2-extensions/identity-oauth2-grant-organization-switch.git</connection>
+        <connection>scm:git:https://github.com/wso2-extensions/identity-oauth2-grant-organization-switch.git
+        </connection>
         <url>https://github.com/wso2-extensions/identity-oauth2-grant-organization-switch.git</url>
         <developerConnection>scm:git:https://github.com/wso2-extensions/identity-oauth2-grant-organization-switch.git
         </developerConnection>
@@ -51,6 +53,11 @@
             <dependency>
                 <groupId>org.wso2.carbon</groupId>
                 <artifactId>org.wso2.carbon.user.core</artifactId>
+                <version>${carbon.kernel.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>org.wso2.carbon</groupId>
+                <artifactId>org.wso2.carbon.user.api</artifactId>
                 <version>${carbon.kernel.version}</version>
             </dependency>
             <dependency>
@@ -197,14 +204,21 @@
         <!-- Carbon kernel version -->
         <carbon.kernel.version>4.7.0-beta9</carbon.kernel.version>
         <carbon.kernel.package.import.version.range>[4.7.0, 5.0.0)</carbon.kernel.package.import.version.range>
+        <carbon.user.api.imp.pkg.version.range>[1.0.1, 2.0.0)</carbon.user.api.imp.pkg.version.range>
 
         <carbon.identity.org.mgt.version>1.0.62</carbon.identity.org.mgt.version>
+        <org.wso2.identity.organization.mgt.imp.pkg.version.range>[1.0.0,2.0.0)
+        </org.wso2.identity.organization.mgt.imp.pkg.version.range>
 
         <carbon.identity.framework.version>5.21.0</carbon.identity.framework.version>
         <carbon.identity.package.import.version.range>[5.20.0, 6.0.0)
         </carbon.identity.package.import.version.range>
         <identity.inbound.auth.oauth.version>6.7.116</identity.inbound.auth.oauth.version>
-        <identity.inbound.auth.oauth.import.version.range>[6.7.116, 6.8.0)</identity.inbound.auth.oauth.import.version.range>
+        <identity.inbound.auth.oauth.import.version.range>[6.7.116, 6.8.0)
+        </identity.inbound.auth.oauth.import.version.range>
+
+        <org.apache.commons.lang.imp.pkg.version.range>[2.6,3)</org.apache.commons.lang.imp.pkg.version.range>
+        <org.apache.commons.logging.imp.pkg.version.range>[1.2,2)</org.apache.commons.logging.imp.pkg.version.range>
 
         <oltu.oauth2.client.version>1.0.0</oltu.oauth2.client.version>
         <oltu.oauth2.client.version.range>[1.0.0,1.0.3)</oltu.oauth2.client.version.range>
